+++ docs/examples/max_concurrency/max_concurrency_asyncio.py
@@ -5,14 +5,14 @@ async def main() -> None:
     async with UnlimitedBucket() as bucket:
         rate_limiter = RateLimiter(bucket, max_concurrency=1)
 
-        with rate_limiter.hold():
+        async with rate_limiter.request():
             print('First request passes')
             try:
-                rate_limiter.hold().__enter__()
+                async with rate_limiter.request(): ...
             except RateLimit:
                 print('Additional concurrent request is rejected')
 
-        with rate_limiter.hold():
+        async with rate_limiter.request():
             print('New request passes when the first one is complete')
 
 run(main())