+++ docs/examples/scheduler/scheduler_trio.py
@@ -1,8 +1,8 @@
 from trio import current_time, open_nursery, run
-from rate_control import Duration, Scheduler, FixedWindowCounter
+from rate_control import Duration, FixedWindowCounter, RateController, Scheduler
 
-async def schedule_print(scheduler: Scheduler, start_time: float) -> None:
-    async with scheduler.schedule():
+async def request_print(controller: RateController, start_time: float) -> None:
+    async with controller.request():
         print(f'Elapsed: {current_time() - start_time :.1f} seconds')
 
 async def main() -> None:
@@ -10,6 +10,6 @@ async def main() -> None:
             Scheduler(bucket) as scheduler, \
             open_nursery() as nursery:
         for _ in range(3):
-            nursery.start_soon(schedule_print, scheduler, current_time())
+            nursery.start_soon(request_print, scheduler, current_time())
 
 run(main)