+++ docs/examples/weighting_requests/weighting_requests_trio.py
@@ -1,22 +1,22 @@
 from trio import run
-from rate_control import Duration, RateLimit, RateLimiter, FixedWindowCounter
+from rate_control import Duration, FixedWindowCounter, RateController, RateLimit, RateLimiter
 
-def normal_request(rate_limiter: RateLimiter) -> None:
-    with rate_limiter.hold():  # weight defaults to 1
+async def normal_request(controller: RateController) -> None:
+    async with controller.request():  # weight defaults to 1
         print('Executed normal request')
 
-def heavy_request(rate_limiter: RateLimiter) -> None:
-    with rate_limiter.hold(2):
+async def heavy_request(rate_limiter: RateLimiter) -> None:
+    async with rate_limiter.request(2):
         print('Executed heavy request')
 
 async def main() -> None:
     async with FixedWindowCounter(capacity=3, duration=Duration.HOUR) as bucket:
         rate_limiter = RateLimiter(bucket)
-        heavy_request(rate_limiter)
+        await heavy_request(rate_limiter)
         try:
-            heavy_request(rate_limiter)
+            await heavy_request(rate_limiter)
         except RateLimit:
             print('Heavy request got rate limited')
-        normal_request(rate_limiter)
+        await normal_request(rate_limiter)
 
 run(main)